refactor: split up setup commands
