+++ src/toolbox/setup/esp8266.ts
@@ -1,0 +1,5 @@
+import { print } from 'gluegun'
+
+export default async function (): Promise<void> {
+  print.info('Setting up esp8266 tools')
+}
