+++ src/toolbox/setup/linux.ts
@@ -1,0 +1,5 @@
+import { print } from 'gluegun'
+
+export default async function (): Promise<void> {
+  print.warning('Linux setup is not currently supported')
+}
