+++ src/toolbox/setup/windows.ts
@@ -1,0 +1,5 @@
+import { print } from 'gluegun'
+
+export default async function (): Promise<void> {
+  print.warning('Windows setup is not currently supported')
+}
